Movigg more files into the listener manager directory.
