+++ source/extensions/listener_managers/listener_manager/listener_impl.cc
@@ -27,11 +27,11 @@
 #include "source/common/network/utility.h"
 #include "source/common/protobuf/utility.h"
 #include "source/common/runtime/runtime_features.h"
+#include "source/extensions/listener_managers/listener_manager/active_raw_udp_listener_config.h"
+#include "source/extensions/listener_managers/listener_manager/filter_chain_manager_impl.h"
 #include "source/extensions/listener_managers/listener_manager/listener_manager_impl.h"
-#include "source/server/active_raw_udp_listener_config.h"
 #include "source/server/configuration_impl.h"
 #include "source/server/drain_manager_impl.h"
-#include "source/server/filter_chain_manager_impl.h"
 #include "source/server/transport_socket_config_impl.h"
 
 #ifdef ENVOY_ENABLE_QUIC
