+++ source/server/BUILD
@@ -424,29 +424,6 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "lds_api_lib",
-    srcs = ["lds_api.cc"],
-    hdrs = ["lds_api.h"],
-    deps = [
-        "//envoy/config:subscription_factory_interface",
-        "//envoy/config:subscription_interface",
-        "//envoy/init:manager_interface",
-        "//envoy/server:listener_manager_interface",
-        "//source/common/common:cleanup_lib",
-        "//source/common/config:api_version_lib",
-        "//source/common/config:subscription_base_interface",
-        "//source/common/config:utility_lib",
-        "//source/common/init:target_lib",
-        "//source/common/protobuf:utility_lib",
-        "@envoy_api//envoy/admin/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/route/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_library(
     name = "api_listener_lib",
     srcs = [
         "api_listener_impl.cc",
@@ -456,8 +433,7 @@ envoy_cc_library(
     ],
     deps = [
         ":configuration_lib",
-        ":filter_chain_manager_lib",
-        ":lds_api_lib",
+        ":factory_context_lib",
         ":transport_socket_config_lib",
         "//envoy/network:connection_interface",
         "//envoy/network:listener_interface",
@@ -483,27 +459,12 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "filter_chain_manager_lib",
-    srcs = ["filter_chain_manager_impl.cc"],
-    hdrs = ["filter_chain_manager_impl.h"],
+    name = "factory_context_lib",
+    srcs = ["factory_context_impl.cc"],
+    hdrs = ["factory_context_impl.h"],
     deps = [
-        ":filter_chain_factory_context_callback",
-        "//envoy/config:typed_metadata_interface",
-        "//envoy/matcher:matcher_interface",
-        "//envoy/network:filter_interface",
+        "//envoy/server:factory_context_interface",
         "//envoy/server:instance_interface",
-        "//envoy/server:listener_manager_interface",
-        "//envoy/server:transport_socket_config_interface",
-        "//source/common/common:empty_string",
-        "//source/common/config:utility_lib",
-        "//source/common/init:manager_lib",
-        "//source/common/matcher:matcher_lib",
-        "//source/common/network:cidr_range_lib",
-        "//source/common/network:lc_trie_lib",
-        "//source/common/network/matching:data_impl_lib",
-        "//source/common/network/matching:inputs_lib",
-        "//source/server:configuration_lib",
-        "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
     ],
 )
 
@@ -559,7 +520,6 @@ envoy_cc_library(
         "abseil_optional",
     ],
     deps = [
-        ":active_raw_udp_listener_config",
         ":api_listener_lib",
         ":configuration_lib",
         ":connection_handler_lib",
@@ -673,25 +633,6 @@ envoy_cc_library(
     hdrs = ["transport_socket_config_impl.h"],
     deps = [
         "//envoy/server:transport_socket_config_interface",
-    ],
-)
-
-envoy_cc_library(
-    name = "active_raw_udp_listener_config",
-    srcs = ["active_raw_udp_listener_config.cc"],
-    hdrs = ["active_raw_udp_listener_config.h"],
-    deps = [
-        ":connection_handler_lib",
-        "//envoy/registry",
-    ],
-)
-
-envoy_cc_library(
-    name = "filter_chain_factory_context_callback",
-    hdrs = ["filter_chain_factory_context_callback.h"],
-    deps = [
-        "//envoy/server:filter_config_interface",
-        "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
     ],
 )
