+++ test/extensions/listener_managers/listener_manager/filter_chain_benchmark_test.cc
@@ -8,7 +8,7 @@
 #include "envoy/protobuf/message_validator.h"
 
 #include "source/common/network/socket_impl.h"
-#include "source/server/filter_chain_manager_impl.h"
+#include "source/extensions/listener_managers/listener_manager/filter_chain_manager_impl.h"
 
 #include "test/benchmark/main.h"
 #include "test/mocks/network/mocks.h"
