+++ test/server/BUILD
@@ -82,7 +82,7 @@ envoy_cc_test(
         "//source/common/network:connection_balancer_lib",
         "//source/common/network:udp_packet_writer_handler_lib",
         "//source/common/stats:stats_lib",
-        "//source/server:active_raw_udp_listener_config",
+        "//source/extensions/listener_managers/listener_manager:active_raw_udp_listener_config",
         "//source/server:connection_handler_lib",
         "//test/mocks/access_log:access_log_mocks",
         "//test/mocks/api:api_mocks",
@@ -246,60 +246,6 @@ envoy_cc_test(
     ],
 )
 
-envoy_cc_test(
-    name = "lds_api_test",
-    srcs = ["lds_api_test.cc"],
-    data = [
-        "//test/config/integration/certs",
-    ],
-    deps = [
-        "//source/common/protobuf:utility_lib",
-        "//source/server:lds_api_lib",
-        "//test/mocks/config:config_mocks",
-        "//test/mocks/init:init_mocks",
-        "//test/mocks/protobuf:protobuf_mocks",
-        "//test/mocks/server:listener_manager_mocks",
-        "//test/mocks/upstream:upstream_mocks",
-        "//test/test_common:environment_lib",
-        "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-        "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
-        "@envoy_api//envoy/service/discovery/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
-    name = "filter_chain_manager_impl_test",
-    srcs = ["filter_chain_manager_impl_test.cc"],
-    data = ["//test/extensions/transport_sockets/tls/test_data:certs"],
-    deps = [
-        ":utility_lib",
-        "//source/common/api:os_sys_calls_lib",
-        "//source/common/config:metadata_lib",
-        "//source/common/network:addr_family_aware_socket_option_lib",
-        "//source/common/network:listen_socket_lib",
-        "//source/common/network:socket_option_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/protobuf",
-        "//source/extensions/filters/network/http_connection_manager:config",
-        "//source/extensions/listener_managers/listener_manager:listener_manager_lib",
-        "//source/extensions/transport_sockets/raw_buffer:config",
-        "//source/extensions/transport_sockets/tls:config",
-        "//source/extensions/transport_sockets/tls:ssl_socket_lib",
-        "//source/server:filter_chain_manager_lib",
-        "//test/mocks/network:network_mocks",
-        "//test/mocks/server:drain_manager_mocks",
-        "//test/mocks/server:factory_context_mocks",
-        "//test/mocks/stream_info:stream_info_mocks",
-        "//test/test_common:environment_lib",
-        "//test/test_common:registry_lib",
-        "//test/test_common:simulated_time_system_lib",
-        "//test/test_common:test_time_lib",
-        "//test/test_common:threadsafe_singleton_injector_lib",
-        "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
-    ],
-)
-
 envoy_cc_fuzz_test(
     name = "server_fuzz_test",
     srcs = ["server_fuzz_test.cc"],
@@ -412,31 +358,6 @@ envoy_cc_test(
         "//test/mocks/thread_local:thread_local_mocks",
         "//test/test_common:utility_lib",
     ],
-)
-
-envoy_cc_benchmark_binary(
-    name = "filter_chain_benchmark_test",
-    srcs = ["filter_chain_benchmark_test.cc"],
-    external_deps = [
-        "benchmark",
-        "googletest",
-    ],
-    deps = [
-        "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
-        "//source/server:filter_chain_manager_lib",
-        "//test/test_common:environment_lib",
-        "//test/mocks/network:network_mocks",
-        "//test/mocks/server:factory_context_mocks",
-        "//test/mocks/stream_info:stream_info_mocks",
-        # tranport socket config registration
-        "//source/extensions/transport_sockets/tls:config",
-    ],
-)
-
-envoy_benchmark_test(
-    name = "filter_chain_benchmark_test_benchmark_test",
-    timeout = "long",
-    benchmark_binary = "filter_chain_benchmark_test",
 )
 
 envoy_cc_benchmark_binary(
