+++ test/server/connection_handler_test.cc
@@ -20,7 +20,7 @@
 #include "source/common/network/udp_listener_impl.h"
 #include "source/common/network/udp_packet_writer_handler_impl.h"
 #include "source/common/network/utility.h"
-#include "source/server/active_raw_udp_listener_config.h"
+#include "source/extensions/listener_managers/listener_manager/active_raw_udp_listener_config.h"
 #include "source/server/connection_handler_impl.h"
 
 #include "test/mocks/access_log/mocks.h"
